refactor(photos,stock): extract components and migrate stock partials
Splits the photos and stock index pages into presentational components and a
usePhotoIndex hook, folds the repeated quantity inputs into a QuantityField,
and moves the stock modals out of the legacy partials/ folder. Refs #81.
