+++ resources/js/pages/stock/components/delete-confirmation.tsx
