+++ resources/js/pages/stock/index.tsx
@@ -1,23 +1,14 @@
 import { PaginationNav } from '@/components/paginationNav';
-import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
-import {
-    Table,
-    TableBody,
-    TableCell,
-    TableHead,
-    TableHeader,
-    TableRow,
-} from '@/components/ui/table';
 import { Searchbar } from '@/features/searchbar';
 import AppLayout from '@/layouts/app-layout';
-import { onSort } from '@/lib/services/filter';
 import { BreadcrumbItem } from '@/types';
-import { Head, Link, router } from '@inertiajs/react';
-import { ArrowUpDown, Diff, Edit2, History, Plus, Trash } from 'lucide-react';
+import { Head, Link } from '@inertiajs/react';
+import { History, Plus } from 'lucide-react';
 import { useState } from 'react';
-import { AlterStockForm } from './partials/alter-stock-form';
-import { DeleteStockableConfirmation } from './partials/delete-confirmation';
+import { AlterStockForm } from './components/alter-stock-form';
+import { DeleteStockableConfirmation } from './components/delete-confirmation';
+import { StockablesTable } from './components/stockables-table';
 
 const breadcrumbs: BreadcrumbItem[] = [
     {
@@ -78,116 +69,13 @@ export default function Stockables({
                         </Button>
                     </div>
                 </div>
-                <Table>
-                    <TableHeader>
-                        <TableRow>
-                            <TableHead className="w-[100px]">
-                                <div className="flex items-center gap-2">
-                                    <button
-                                        onClick={() =>
-                                            onSort('id', 'stockables.index')
-                                        }
-                                    >
-                                        <ArrowUpDown className="h-4 w-4" />
-                                    </button>
-                                    #
-                                </div>
-                            </TableHead>
-                            <TableHead>
-                                <div className="flex items-center gap-2">
-                                    <button
-                                        onClick={() =>
-                                            onSort('name', 'stockables.index')
-                                        }
-                                    >
-                                        <ArrowUpDown className="h-4 w-4" />
-                                    </button>
-                                    Stockeable
-                                </div>
-                            </TableHead>
-                            <TableHead>
-                                <div className="flex items-center gap-2">
-                                    <button
-                                        onClick={() =>
-                                            onSort(
-                                                'quantity',
-                                                'stockables.index',
-                                            )
-                                        }
-                                    >
-                                        <ArrowUpDown className="h-4 w-4" />
-                                    </button>
-                                    Cantidad
-                                </div>
-                            </TableHead>
-                            <TableHead>
-                                <div className="flex items-center gap-2">
-                                    Productos
-                                </div>
-                            </TableHead>
-                            <TableHead>Acciones</TableHead>
-                        </TableRow>
-                    </TableHeader>
-                    <TableBody>
-                        {stockables.data.map((stockable) => (
-                            <TableRow key={stockable.id}>
-                                <TableCell className="font-medium">
-                                    {stockable.id}
-                                </TableCell>
-                                <TableCell>{stockable.name}</TableCell>
-                                <TableCell>
-                                    {stockable.quantity < stockable.alert_at ? (
-                                        <Badge variant="destructive">
-                                            {`${stockable.quantity} (${stockable.unit})`}
-                                        </Badge>
-                                    ) : (
-                                        `${stockable.quantity} (${stockable.unit})`
-                                    )}
-                                </TableCell>
-                                <TableCell>
-                                    {stockable.products
-                                        .map((product) => product.name)
-                                        .join(', ')}
-                                </TableCell>
-                                <TableCell className="flex gap-2">
-                                    <Button
-                                        size={'sm'}
-                                        variant={'warning'}
-                                        onClick={() =>
-                                            router.visit(
-                                                route('stockables.edit', {
-                                                    stockable: stockable.id,
-                                                }),
-                                            )
-                                        }
-                                    >
-                                        <Edit2 />
-                                    </Button>
-                                    <Button
-                                        size={'sm'}
-                                        variant={'destructive'}
-                                        onClick={() =>
-                                            setShowDeleteStockableConfirmation(
-                                                stockable,
-                                            )
-                                        }
-                                    >
-                                        <Trash />
-                                    </Button>
-                                    <Button
-                                        size={'sm'}
-                                        variant="secondary"
-                                        onClick={() =>
-                                            setShowAddStockModal(stockable)
-                                        }
-                                    >
-                                        <Diff />
-                                    </Button>
-                                </TableCell>
-                            </TableRow>
-                        ))}
-                    </TableBody>
-                </Table>
+
+                <StockablesTable
+                    stockables={stockables.data}
+                    onDelete={setShowDeleteStockableConfirmation}
+                    onAlterStock={setShowAddStockModal}
+                />
+
                 <PaginationNav links={stockables.meta.links} />
             </section>
         </AppLayout>
